# api/openapi.yaml
swagger: "2.0"
info:
  title: Configuration Management
  version: 0.0.1
  description: |
    REST API for managing system resources with idempotent semantics. Includes support for
    optimistic concurrency using ETags.
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - http
  - https
basePath: "/api/v1/"
...
paths:
...
  /command:
    post:
      summary: Execute a command on the target system
      description: |
        Executes a command on the target system and returns the result. Commands are
        executed with a configurable timeout and expected exit codes.
      operationId: executeCommand
      tags:
        - Command
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: command
          required: true
          schema:
            $ref: "#/definitions/CommandRequest"
      responses:
        200:
          description: Command executed successfully
          schema:
            $ref: "#/definitions/CommandResponse"
        400:
          description: Invalid request or malformed command
          schema:
            $ref: "#/responses/ErrorResponse"
        408:
          description: Command execution timeout
          schema:
            $ref: "#/responses/ErrorResponse"
        500:
          description: Internal server error during command execution
          schema:
            $ref: "#/responses/ErrorResponse"
...
  /directories:
    get:
      summary: Retrieve the current state and properties of a directory
      description: |
        Fetches properties about the specified directory, including its permissions and
        ownership.
      operationId: getDirectoryProperties
      tags:
        - Directories
      parameters:
        - $ref: "#/parameters/DirectoryPath"
      responses:
        200:
          $ref: "#/responses/DirectoryPropertiesResponse"
        400:
          description: Invalid request or missing fields
          schema:
            $ref: "#/responses/ErrorResponse"
        404:
          description: Directory not found
          schema:
            $ref: "#/responses/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/responses/ErrorResponse"
    put:
      summary: Create or update properties for a directory
      description: |
        Creates a new directory if it does not exist or updates properties (mode, owner,
        group) if the directory exists. Creates parent directories as needed.
      operationId: putDirectory
      tags:
        - Directories
      parameters:
        - $ref: "#/parameters/DirectoryPath"
        - $ref: "#/parameters/IfMatch"
        - in: body
          name: properties
          required: true
          schema:
            $ref: "#/definitions/DirectoryProperties"
      responses:
        201:
          description: Directory created and properties were applied
          headers:
            ETag:
              type: string
        204:
          description: Directory already existed and properties were applied
          headers:
            ETag:
              type: string
        400:
          description: Invalid request
          schema:
            $ref: "#/responses/ErrorResponse"
        409:
          description: ETag mismatch
          schema:
            $ref: "#/responses/ErrorResponse"
        412:
          description: Precondition failed
          schema:
            $ref: "#/responses/ErrorResponse"
        428:
          description: Missing If-Match header
          schema:
            $ref: "#/responses/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/responses/ErrorResponse"
    delete:
      summary: Delete an existing directory
      description: Removes the specified directory from the system
      operationId: deleteDirectory
      tags:
        - Directories
      parameters:
        - $ref: "#/parameters/DirectoryPath"
        - $ref: "#/parameters/IfMatch"
      responses:
        204:
          description: Directory deleted
        400:
          description: Invalid request
          schema:
            $ref: "#/responses/ErrorResponse"
        403:
          description: Forbidden
          schema:
            $ref: "#/responses/ErrorResponse"
        409:
          description: ETag mismatch
          schema:
            $ref: "#/responses/ErrorResponse"
        428:
          description: Missing If-Match header
          schema:
            $ref: "#/responses/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/responses/ErrorResponse"
parameters:
  IfMatch:
    name: If-Match
    in: header
    type: string
    required: false
    description: ETag value for optimistic concurrency control
...
  DirectoryPath:
    name: path
    in: query
    required: true
    type: string
    description: Absolute directory path on the target system
responses:
  ErrorResponse:
    description: Standard error response
    schema:
      $ref: "#/definitions/Error"
...
  DirectoryPropertiesResponse:
    description: Directory properties with metadata
    headers:
      ETag:
        type: string
        description: ETag for optimistic concurrency control
    schema:
      $ref: "#/definitions/DirectoryProperties"
definitions:
  Error:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
        description: Human-readable error message
      details:
        type: string
        description: Additional error details for debugging
  CommandRequest:
    type: object
    properties:
      command:
        type: string
        description: The command to execute
        example: "ls -la /etc"
      expected_exit_codes:
        type: array
        items:
          type: integer
        default: [0]
        description: Expected exit codes for success (default [0])
  CommandResponse:
    type: object
    properties:
      exit_code:
        type: integer
        description: Command exit code
      stdout:
        type: string
        description: Standard output from command
      stderr:
        type: string
        description: Standard error from command
      success:
        type: boolean
        description: Whether command execution was considered successful
...
  DirectoryProperties:
    type: object
    properties:
      mode:
        type: string
        pattern: '^[0-7]{3,4}$'
        description: Directory permissions in octal format (e.g., "0755")
        example: "0755"
      owner:
        type: string
      group:
        type: string